Change debug variable type to uint8_t

+++ src/debug/debug.c
@@ -1,4 +1,4 @@
 #include "debug.h"
 
-int debug = 0;
+uint8_t debug = 0;
 Debug dbg = { 0, 0, 0, 0 };
